Split MDS DOI helpers off the protocol base controller.
Replace kitchen-sink DoiSupport with DoiLookup/DoiWriter includes only where
needed, move minting to Mds::DoiMinter, use authorize! only in upsert, and
drop the Rails.env.test? uniqueness bypass.
